refactor(about): separate into smaller components
