+++ src/AdminDashboard/About/AboutList.js
@@ -1,0 +1,64 @@
+import React, { Component } from "react"
+import ReactWinJS from 'react-winjs'
+import PropTypes from 'prop-types'
+let WinJS = require('winjs')
+
+export default class AboutList extends Component {
+
+    constructor(props) {
+        super(props)
+        this.state = {
+            layout: { type: WinJS.UI.ListLayout }
+        }
+    }
+
+    itemRenderer = ReactWinJS.reactRenderer((item) => {
+        return (
+            <div style={{ padding: '14px', width: '100%' }}>
+                {item.data.title}
+            </div>
+        )
+    })
+
+
+    handleSelectionChanged = (eventObject) => {
+        let listView = eventObject.currentTarget.winControl
+        let index = listView.selection.getIndices()
+        this.props.onNavigate([this.props.location, index[0]])
+    }
+
+    handleContentAnimating(eventObject) {
+        // Disable ListView's entrance animation
+        if (eventObject.detail.type === 'entrance') {
+            eventObject.preventDefault()
+        }
+    }
+
+    render() {
+        return (
+            <div className="listPane" style={{ height: '100%', width: this.props.itemListPaneWidth, display: 'inline-block', verticalAlign: 'top' }}>
+                <ReactWinJS.ListView
+                    ref="listView"
+                    className="contentListView win-selectionstylefilled"
+                    style={{ height: 'calc(100% - 48px)' }}
+                    itemDataSource={this.props.itemList.dataSource}
+                    itemTemplate={this.itemRenderer}
+                    layout={this.state.layout}
+                    selectionMode="single"
+                    tapBehavior="directSelect" 
+                    onContentAnimating={this.handleContentAnimating}
+                    onSelectionChanged={this.handleSelectionChanged}
+                />
+            </div>
+        )
+    }
+}
+AboutList.propTypes = {
+    itemListPaneWidth: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number
+    ]).isRequired,
+    itemList: PropTypes.object.isRequired,
+    location: PropTypes.array.isRequired,
+    onNavigate: PropTypes.func.isRequired
+}
